Integrate new contract generation
